search-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
transformers>=4.35.0
faiss-cpu>=1.7.4
pillow>=10.0.0
numpy>=1.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
requests>=2.31.0
# Deployment specific
gunicorn==21.2.0 